+++ src/configs/ts.ts
@@ -267,7 +267,10 @@ export const tsUnConfig: UnConfigFn<
           },
         },
       )
-      .disableAnyRule(['no-unused-vars', 'no-use-before-define', 'no-shadow', 'no-redeclare']);
+      .disableAnyRule('', 'no-unused-vars')
+      .disableAnyRule('', 'no-use-before-define')
+      .disableAnyRule('', 'no-shadow')
+      .disableAnyRule('', 'no-redeclare');
 
     return configBuilderSetup;
   };
@@ -386,7 +389,7 @@ export const tsUnConfig: UnConfigFn<
     // .addRule('no-invalid-this', OFF, [], {overrideBaseRule: true}) // 👍
     .addRule('no-loop-func', ERROR, [], {overrideBaseRule: true})
     // .addRule('no-magic-numbers', OFF, [], {overrideBaseRule: true})
-    .disableAnyRule('no-redeclare')
+    .disableAnyRule('', 'no-redeclare')
     // .addRule('no-redeclare', OFF) // 👍
     .addRule('no-require-imports', OFF)
     // .addRule('no-restricted-imports', OFF, [], {overrideBaseRule: true})
@@ -403,7 +406,8 @@ export const tsUnConfig: UnConfigFn<
     // .addRule('prefer-enum-initializers', OFF)
     // .addRule('typedef', OFF)
     // 🟢 Disable conflicting rules
-    .disableAnyRule(['no-useless-constructor', 'dot-notation'])
+    .disableAnyRule('', 'no-useless-constructor')
+    .disableAnyRule('', 'dot-notation')
     .addOverrides();
 
   // CONFIG TYPE AWARE
@@ -494,7 +498,7 @@ export const tsUnConfig: UnConfigFn<
     .addRule('no-unsafe-return', noUnsafeRulesSeverity)
     // .addRule('no-unsafe-type-assertion', OFF)
     // .addRule('no-unsafe-unary-minus', ERROR)
-    .disableAnyRule('no-throw-literal') // Note: has different name
+    .disableAnyRule('', 'no-throw-literal') // Note: has different name
     .addRule('only-throw-error', ERROR, [
       {
         allowRethrowing: true,
@@ -516,7 +520,7 @@ export const tsUnConfig: UnConfigFn<
     })
     // .addRule('non-nullable-type-assertion-style', ERROR)
     // .addRule('prefer-find', ERROR)
-    .disableAnyRule('unicorn/prefer-includes')
+    .disableAnyRule('unicorn', 'prefer-includes')
     // .addRule('prefer-includes', ERROR)
     .addRule('prefer-nullish-coalescing', OFF)
     // .addRule('prefer-optional-chain', ERROR)
@@ -529,14 +533,14 @@ export const tsUnConfig: UnConfigFn<
     .addRule('prefer-destructuring', ERROR, RULE_PREFER_DESTRUCTURING_OPTIONS, {
       overrideBaseRule: true,
     })
-    .disableAnyRule('unicorn/prefer-array-find') // Note: in Unicorn
+    .disableAnyRule('unicorn', 'prefer-array-find') // Note: in Unicorn
     .addRule('prefer-readonly', ERROR)
     // .addRule('prefer-readonly-parameter-types', OFF)
     // .addRule('promise-function-async', OFF)
     // .addRule('related-getter-setter-pairs', ERROR)
     // .addRule('require-array-sort-compare', OFF)
     // Note: has different name. Also note that the original rule is deprecated and not included in this config, but we disable it anyway just for safety
-    .disableAnyRule('no-return-await') // Disabled by default since v8
+    .disableAnyRule('', 'no-return-await') // Disabled by default since v8
     .addRule('return-await', ERROR, ['always'])
     // .addRule('strict-boolean-expressions', OFF)
     .addRule('switch-exhaustiveness-check', ERROR)
@@ -547,31 +551,29 @@ export const tsUnConfig: UnConfigFn<
     ?.addConfig('ts/disable-handled-by-ts-compiler-rules', {
       files: [...TS_FILES_DEFAULT, ...filesNONTypeAware, ...filesTypeAware],
     })
-    .disableAnyRule([
-      'constructor-super',
-      'getter-return',
-      'no-const-assign',
-      'no-dupe-args',
-      'no-dupe-class-members',
-      'no-dupe-keys',
-      'no-func-assign',
-    ])
+    .disableAnyRule('', 'constructor-super')
+    .disableAnyRule('', 'getter-return')
+    .disableAnyRule('', 'no-const-assign')
+    .disableAnyRule('', 'no-dupe-args')
+    .disableAnyRule('', 'no-dupe-class-members')
+    .disableAnyRule('', 'no-dupe-keys')
+    .disableAnyRule('', 'no-func-assign')
     // "Note that the compiler will not catch the Object.assign() case. Thus, if you use Object.assign() in your codebase, this rule will still provide some value." - https://eslint.org/docs/latest/rules/no-import-assign#handled_by_typescript
     // .addRule('no-import-assign', OFF)
     // "Note that, technically, TypeScript will only catch this if you have the strict or noImplicitThis flags enabled. These are enabled in most TypeScript projects, since they are considered to be best practice." - https://eslint.org/docs/latest/rules/no-invalid-this#rule-details
     // .addRule('no-invalid-this', OFF)
-    .disableAnyRule([
-      'no-new-native-nonconstructor', // successor of no-new-symbol
-      'no-obj-calls',
-    ])
+    .disableAnyRule('', 'no-new-native-nonconstructor') // successor of no-new-symbol
+    .disableAnyRule('', 'no-obj-calls')
     // "Note that while TypeScript will catch let redeclares and const redeclares, it will not catch var redeclares. Thus, if you use the legacy var keyword in your TypeScript codebase, this rule will still provide some value." - https://eslint.org/docs/latest/rules/no-redeclare#handled_by_typescript
     // .addRule('no-redeclare', OFF)
-    .disableAnyRule(['no-setter-return', 'no-this-before-super', 'no-undef'])
+    .disableAnyRule('', 'no-setter-return')
+    .disableAnyRule('', 'no-this-before-super')
+    .disableAnyRule('', 'no-undef')
     // "TypeScript must be configured with allowUnreachableCode: false for it to consider unreachable code an error." - https://eslint.org/docs/latest/rules/no-unreachable#handled_by_typescript
     // .addRule('no-unreachable', OFF)
-    .disableAnyRule('no-unsafe-negation')
+    .disableAnyRule('', 'no-unsafe-negation')
     // Does not work correctly when type-only imports are present because you can't combine such an import with a default import.
-    .disableAnyRule('no-duplicate-imports');
+    .disableAnyRule('', 'no-duplicate-imports');
 
   const configBuilderDts = createConfigBuilder(context, {}, '@typescript-eslint');
   configBuilderDts
@@ -586,14 +588,12 @@ export const tsUnConfig: UnConfigFn<
     .addRule('no-unnecessary-type-parameters', OFF)
     .addRule('no-unused-vars', OFF)
     .addRule('no-use-before-define', OFF)
-    .disableAnyRule([
-      'import/newline-after-import',
-      'import/no-default-export',
-      'vars-on-top',
-      'no-var',
-      'sonarjs/no-redundant-optional',
-      'no-duplicate-imports',
-    ]);
+    .disableAnyRule('import', 'newline-after-import')
+    .disableAnyRule('import', 'no-default-export')
+    .disableAnyRule('', 'vars-on-top')
+    .disableAnyRule('', 'no-var')
+    .disableAnyRule('sonarjs', 'no-redundant-optional')
+    .disableAnyRule('', 'no-duplicate-imports');
 
   const configBuilderNoTypeAssertions = createConfigBuilder(
     context,