+++ src/configs/unused-imports.ts
@@ -59,11 +59,9 @@ export const unusedImportsUnConfig: UnConfigFn<'unusedImports'> = (context) => {
 
   configBuilderNoUnusedVars
     ?.addConfig(['unused-imports/no-unused-vars', {includeDefaultFilesAndIgnores: true}])
-    .disableAnyRule([
-      'no-unused-vars',
-      'sonarjs/no-unused-vars',
-      '@typescript-eslint/no-unused-vars',
-    ])
+    .disableAnyRule('', 'no-unused-vars')
+    .disableAnyRule('sonarjs', 'no-unused-vars')
+    .disableAnyRule('@typescript-eslint', 'no-unused-vars')
     .addRule('no-unused-vars', ERROR, ruleOptions === undefined ? [] : [ruleOptions]);
 
   return {