+++ src/configs/vue.ts
@@ -170,6 +170,7 @@ export const vueUnConfig: UnConfigFn<'vue'> = async (context) => {
     eslintParserVue,
     isPiniaPackageInstalled,
     nuxtPackageInfo,
+    {parser: typescriptEslintParser},
   ] = await Promise.all([
     interopDefault(import('eslint-merge-processors')),
     interopDefault(import('eslint-processor-vue-blocks')),
@@ -179,6 +180,7 @@ export const vueUnConfig: UnConfigFn<'vue'> = async (context) => {
     interopDefault(import('vue-eslint-parser')),
     doesPackageExist('pinia'),
     fetchPackageInfo('nuxt'),
+    interopDefault(import('typescript-eslint')),
   ]);
 
   const isTypescriptEnabled = context.configsMeta.ts.enabled;
@@ -266,11 +268,7 @@ export const vueUnConfig: UnConfigFn<'vue'> = async (context) => {
       globals: globals.browser,
       parser: eslintParserVue,
       parserOptions: {
-        ecmaFeatures: {
-          jsx: true,
-        },
-        extraFileExtensions: ['.vue'],
-        parser: isTypescriptEnabled ? '@typescript-eslint/parser' : undefined,
+        parser: isTypescriptEnabled ? typescriptEslintParser : undefined,
         sourceType: 'module' as const,
       },
     },
@@ -627,13 +625,11 @@ export const vueUnConfig: UnConfigFn<'vue'> = async (context) => {
     // .addRule('space-unary-ops', OFF)
     // .addRule('template-curly-spacing', OFF)
     // 🔵 Not working great in Vue files
-    .disableAnyRule([
-      '@typescript-eslint/prefer-function-type',
-      '@typescript-eslint/unified-signatures',
-      'import/first', // May be wrong if multiple <script> blocks are present
-      'import/no-default-export',
-      'no-useless-assignment', // False positives in script setup
-    ])
+    .disableAnyRule('@typescript-eslint', 'prefer-function-type')
+    .disableAnyRule('@typescript-eslint', 'unified-signatures')
+    .disableAnyRule('import', 'first') // May be wrong if multiple <script> blocks are present
+    .disableAnyRule('import', 'no-default-export')
+    .disableAnyRule('', 'no-useless-assignment') // False positives in script setup
     .addOverrides();
 
   configBuilder
@@ -672,7 +668,7 @@ export const vueUnConfig: UnConfigFn<'vue'> = async (context) => {
         .flat()
         .filter((v) => typeof v === 'string'),
     })
-    .disableAnyRule('import/no-default-export');
+    .disableAnyRule('import', 'no-default-export');
 
   const configBuilderA11y = createConfigBuilder(context, configA11y, 'vuejs-accessibility');
   configBuilderA11y